+++ openapi/models/operations/getaward.go
@@ -10,12 +10,12 @@ import (
 
 type GetAwardRequest struct {
 	// id
-	ID int64 `pathParam:"style=simple,explode=false,name=id"`
+	ID float64 `pathParam:"style=simple,explode=false,name=id"`
 }
 
-func (o *GetAwardRequest) GetID() int64 {
+func (o *GetAwardRequest) GetID() float64 {
 	if o == nil {
-		return 0
+		return 0.0
 	}
 	return o.ID
 }
