+++ openapi/models/operations/getawardcategoryextended.go
@@ -10,12 +10,12 @@ import (
 
 type GetAwardCategoryExtendedRequest struct {
 	// id
-	ID int64 `pathParam:"style=simple,explode=false,name=id"`
+	ID float64 `pathParam:"style=simple,explode=false,name=id"`
 }
 
-func (o *GetAwardCategoryExtendedRequest) GetID() int64 {
+func (o *GetAwardCategoryExtendedRequest) GetID() float64 {
 	if o == nil {
-		return 0
+		return 0.0
 	}
 	return o.ID
 }
