+++ openapi/models/operations/getawardextended.go
@@ -10,12 +10,12 @@ import (
 
 type GetAwardExtendedRequest struct {
 	// id
-	ID int64 `pathParam:"style=simple,explode=false,name=id"`
+	ID float64 `pathParam:"style=simple,explode=false,name=id"`
 }
 
-func (o *GetAwardExtendedRequest) GetID() int64 {
+func (o *GetAwardExtendedRequest) GetID() float64 {
 	if o == nil {
-		return 0
+		return 0.0
 	}
 	return o.ID
 }
