+++ openapi/models/operations/getgenrebase.go
@@ -10,12 +10,12 @@ import (
 
 type GetGenreBaseRequest struct {
 	// id
-	ID int64 `pathParam:"style=simple,explode=false,name=id"`
+	ID float64 `pathParam:"style=simple,explode=false,name=id"`
 }
 
-func (o *GetGenreBaseRequest) GetID() int64 {
+func (o *GetGenreBaseRequest) GetID() float64 {
 	if o == nil {
-		return 0
+		return 0.0
 	}
 	return o.ID
 }
