+++ openapi/models/operations/getmoviebase.go
@@ -10,12 +10,12 @@ import (
 
 type GetMovieBaseRequest struct {
 	// id
-	ID int64 `pathParam:"style=simple,explode=false,name=id"`
+	ID float64 `pathParam:"style=simple,explode=false,name=id"`
 }
 
-func (o *GetMovieBaseRequest) GetID() int64 {
+func (o *GetMovieBaseRequest) GetID() float64 {
 	if o == nil {
-		return 0
+		return 0.0
 	}
 	return o.ID
 }
