+++ openapi/models/operations/getseasontranslation.go
@@ -10,14 +10,14 @@ import (
 
 type GetSeasonTranslationRequest struct {
 	// id
-	ID int64 `pathParam:"style=simple,explode=false,name=id"`
+	ID float64 `pathParam:"style=simple,explode=false,name=id"`
 	// language
 	Language string `pathParam:"style=simple,explode=false,name=language"`
 }
 
-func (o *GetSeasonTranslationRequest) GetID() int64 {
+func (o *GetSeasonTranslationRequest) GetID() float64 {
 	if o == nil {
-		return 0
+		return 0.0
 	}
 	return o.ID
 }
